refactor(server): split HttpApi exercise harness
